refactor: separate the constructor parser
