+++ src/parse/declaration/entity.rs
@@ -1,11 +1,8 @@
 use crate::{
     ast::*,
     parse::{
-        expression::call::{block_parser, call_arg_parser},
-        ty::{
-            class_type_params_parser, function::function_params_parser,
-            type_bounds_parser, type_parser,
-        },
+        expression::call::block_parser,
+        ty::{class_type_params_parser, type_bounds_parser, type_parser},
     },
 };
 use chumsky::prelude::*;
@@ -19,10 +16,7 @@ pub fn entity_parser(
     modifier_parser()
         .repeated()
         .or_not()
-        .then(
-            entity_kind_parser(stmt_parser.clone(), expr_parser.clone())
-                .padded(),
-        )
+        .then(entity_kind_parser().padded())
         .then(text::ident().padded().or_not())
         .then(
             class_type_params_parser(expr_parser.clone())
@@ -62,8 +56,6 @@ pub fn entity_parser(
 }
 
 pub fn entity_kind_parser(
-    stmt_parser: impl Parser<char, Statement, Error = Simple<char>>,
-    expr_parser: impl Parser<char, Expression, Error = Simple<char>> + Clone,
 ) -> impl Parser<char, EntityDeclarationKind, Error = Simple<char>> {
     choice((
         just("enum")
@@ -77,8 +69,6 @@ pub fn entity_kind_parser(
             .then(just("object"))
             .to(EntityDeclarationKind::CompanionObject),
         just("object").to(EntityDeclarationKind::Object),
-        secondary_constructor_parser(stmt_parser, expr_parser)
-            .map(EntityDeclarationKind::SecondaryConstructor),
     ))
 }
 
@@ -138,50 +128,5 @@ pub fn constructor_param_parser(
                     ty,
                 },
             }
-        })
-}
-
-pub fn secondary_constructor_parser(
-    stmt_parser: impl Parser<char, Statement, Error = Simple<char>>,
-    expr_parser: impl Parser<char, Expression, Error = Simple<char>> + Clone,
-) -> impl Parser<char, ConstructorDeclaration, Error = Simple<char>> {
-    modifier_parser()
-        .repeated()
-        .or_not()
-        .then_ignore(just("constructor").padded())
-        .then(
-            function_params_parser(expr_parser.clone())
-                .delimited_by(just('(').padded(), just(')').padded()),
-        )
-        .then(constructor_delegate_parser(expr_parser).or_not())
-        .then(block_parser(stmt_parser).or_not())
-        .map(
-            |(((modifiers, params), delegate), body)| ConstructorDeclaration {
-                modifiers: modifiers.unwrap_or_default(),
-                params,
-                delegate,
-                body,
-            },
-        )
-}
-
-pub fn constructor_delegate_parser(
-    expr_parser: impl Parser<char, Expression, Error = Simple<char>>,
-) -> impl Parser<char, ConstructorDelegate, Error = Simple<char>> {
-    just(':')
-        .padded()
-        .ignore_then(choice((
-            just("this").to(ConstructorDelegateKind::This),
-            just("super").to(ConstructorDelegateKind::Super),
-        )))
-        .then(
-            call_arg_parser(expr_parser)
-                .separated_by(just(',').padded())
-                .or_not()
-                .delimited_by(just('(').padded(), just(')').padded()),
-        )
-        .map(|(kind, args)| ConstructorDelegate {
-            kind,
-            args: args.unwrap_or_default(),
         })
 }
